SNAKE GAME - PYTHON
Embark on a timeless journey with the classic Snake game in Python, where simplicity meets charm. Navigate through a retro-inspired world as your snake hungrily devours delectable treats, each morsel contributing to its growing size. However, tread cautiously, for a single misstep, be it a collision with your own tail or the unforgiving walls, heralds the end of your adventure with a solemn declaration: "Game Over!"